find one signal

+++ signal/signal.go
@@ -3,7 +3,7 @@ package signal
 import "time"
 
 type Signal struct {
-	ID              int                    `json:"_id"`
+	ID              int64                  `json:"_id"`
 	ServiceID       string                 `json:"service_id"`
 	ServiceRuleID   string                 `json:"service_rule_id"`
 	OutgoingPayload map[string]interface{} `json:"outgoing_payload"`
